[mm] vmstat: fix overflow in mod_zone_page_state()
Message-id: <1462810725-31663-2-git-send-email-brueckner@redhat.com>
Patchwork-id: 143728
O-Subject: [RHEL7.3 PATCH 1/2] [mm] mm/vmstat: fix overflow in mod_zone_page_state()
Bugzilla: 1334242
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
commit 6cdb18ad98a49f7e9b95d538a0614cde827404b8
Author: Heiko Carstens <heiko.carstens@de.ibm.com>
Commit: Linus Torvalds <torvalds@linux-foundation.org>

    mm/vmstat: fix overflow in mod_zone_page_state()

    mod_zone_page_state() takes a "delta" integer argument.  delta contains
    the number of pages that should be added or subtracted from a struct
    zone's vm_stat field.

    If a zone is larger than 8TB this will cause overflows.  E.g.  for a
    zone with a size slightly larger than 8TB the line

        mod_zone_page_state(zone, NR_ALLOC_BATCH, zone->managed_pages);

    in mm/page_alloc.c:free_area_init_core() will result in a negative
    result for the NR_ALLOC_BATCH entry within the zone's vm_stat, since 8TB
    contain 0x8xxxxxxx pages which will be sign extended to a negative
    value.

    Fix this by changing the delta argument to long type.

    This could fix an early boot problem seen on s390, where we have a 9TB
    system with only one node.  ZONE_DMA contains 2GB and ZONE_NORMAL the
    rest.  The system is trying to allocate a GFP_DMA page but ZONE_DMA is
    completely empty, so it tries to reclaim pages in an endless loop.

    This was seen on a heavily patched 3.10 kernel.  One possible
    explaination seem to be the overflows caused by mod_zone_page_state().
    Unfortunately I did not have the chance to verify that this patch
    actually fixes the problem, since I don't have access to the system
    right now.  However the overflow problem does exist anyway.

    Given the description that a system with slightly less than 8TB does
    work, this seems to be a candidate for the observed problem.

    Cc: Christoph Lameter <cl@linux.com>

Bugzilla
--------
BZ 1334242
https://bugzilla.redhat.com/show_bug.cgi?id=1334242

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version v4.4-rc8
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=6cdb18ad98a49f7e9b95d538a0614cde827404b8

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=10985824

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

+++ include/linux/vmstat.h
@@ -180,11 +180,11 @@ extern void zone_statistics(struct zone *, struct zone *, gfp_t gfp);
 #define sub_zone_page_state(__z, __i, __d) mod_zone_page_state(__z, __i, -(__d))
 
 #ifdef CONFIG_SMP
-void __mod_zone_page_state(struct zone *, enum zone_stat_item item, int);
+void __mod_zone_page_state(struct zone *, enum zone_stat_item item, long);
 void __inc_zone_page_state(struct page *, enum zone_stat_item);
 void __dec_zone_page_state(struct page *, enum zone_stat_item);
 
-void mod_zone_page_state(struct zone *, enum zone_stat_item, int);
+void mod_zone_page_state(struct zone *, enum zone_stat_item, long);
 void inc_zone_page_state(struct page *, enum zone_stat_item);
 void dec_zone_page_state(struct page *, enum zone_stat_item);
 
@@ -221,7 +221,7 @@ static inline void __dec_zone_state(struct zone *zone, enum zone_stat_item item)
  * The functions directly modify the zone and global counters.
  */
 static inline void __mod_zone_page_state(struct zone *zone,
-			enum zone_stat_item item, int delta)
+			enum zone_stat_item item, long delta)
 {
 	zone_page_state_add(delta, zone, item);
 }
